PLATFORM-2362: Add proper support for incoming X-Trace-Id
